Norm data
Norm data of 31 subjects. The description of the subjects is in the excel file. This file will be needed in the script_201811.m
c3d: data are not filtered, the threshold of the forceplate is set to 5N, there is a general event on left or right to define the step that will be checked (the first step after the general event)
In the excel file are given which trial and which forceplate are used for which foot (e.g. Tr 12, FP 1)